// src/features/AutoTestsGenerateForm/model/FormMain/FormMain.types.ts
export interface IFormAutoTestsInput {
	url: string
	stack: string
	title_check: string[]
	console_check: string[]
	manual: string
	title_1: string
	title_2: string
	title_3: string
	steps_1: string
	steps_2: string
	steps_3: string
	allure_report: string[]
	attachments: string[]
	source_code: string
	ci: string
	browser_type: string
	browsers_hub: string
	notification: string[]
	test_management: string[]
}